Jenkins: Do not checkout PR branch on layer repos
As the PR branch within jenkins does not contain the actual
branch name, this will not lead to the expected result when trying to
checkout the same branch in the layer repos. Instead, use the real
branch name for these checkout operations.

Additionally, we only try to checkout in repos containing the "iris"
string.