Enable adding patch files to KAS repository
Previously, files for patching one of the yocto layer repositories had
to added within another yocto layer repository. This commit prepares the
KAS repository for adding patches directly within the KAS repository.
Existing tooling as been adapted for the inclusion of the "this"
repository.

As a side effect, the KAS repository is now also added to the yocto
BBLAYERS. This however does not seem to have a negative effect.