[DEVOPS-296] Reduce number of gtest runs
Currently our QEMU setup does not support KVM and only runs with
user networking, which drastically reduces the performance. This is
something that will change in the long run (as soon as we quit using
AWS codebuild), however for now we will reduce the number of re-runs, as
to reduce the time spent in the QEMU test step.