[DEVOPS-602] Workaround for gitlab constraints
Currently, only maintainers can run pipelines on protected branches, as
the permission for running CI jobs is bound to push permissions
(https://gitlab.com/gitlab-org/gitlab/-/issues/334413)
This is a constraint when trying to trigger a downstream pipeline using
the develop branch of a project
(https://gitlab.com/gitlab-org/gitlab/-/issues/299433).

As a workaround, we use a separate, unprotected branch, which is
automatically reset to the develop branch revision when running a
upstream pipeline.