[DEVOPS-637] gitlab-ci: Handle MR to non-default branch
meta-layer pipeline:
Running a build pipeline on release merge makes sense, as this will
check the correct configuration of this particular layer before a
release.

iris-kas pipeline:
During a release within the iris-kas repository only the RC-x runs are
of interest, as the release (tag) pipeline fundamentally differs from the
development pipeline. As such, we skip the merge request by adding an
empty job