Remove needs keyword due to issues with build
Apperantly there is a bug in GitLab which will cause the pipeline to
continue to run, if a previous job failed but no "need" has been
specified for these jobs. To prevent any mishaps, we will remove the
need keyword feature for now, which will result in waiting for the
previous stage to completely finish before running qemu tests.