Update KAS_PREMIRRORS to include iris repos
Previously, KAS_PREMIRRORS would exclude any iris-specific repositories,
which made sense at the time as the primary repositories in question
were hosted on github (meta-iris-base). As these repositories have moved
to gitlab now and are only mirrored to github, it makes sense to use a
premirror for these as well.