Populate sstate-cache on meta layer merge
Previously, all iris-kas child pipelines triggered by a meta layer parent
pipeline would run on the gitlab-pipeline branch in the iris-kas repo.
This is to work around permission errors when non-maintainers trigger a
child pipeline on the protected iris-kas default branch.
However, this means that pushes to meta layer repos default branch
will not populate the sstate cache (which is only done when running on
iris-kas default branch).

Since only maintainers can merge on the meta layer repos and should
also have maintainer permissions on iris-kas, we can cause runs on the
meta layers default branch to use the iris-kas default branch in the
child pipeline.