[DEVOPS-600] Fix creating gtest result artifacts on failure
Previously, failing the ssh command step would cause the whole job to
fail early, thus the results of the gtest run would never be added as
artifact.

Move the scp command step into after_script, which is always executed,
even on failure.

Also, do not fail gtest early, so that all failures are accounted for in the
resulting report.