Make uploading built SDKs to registries optional
We want to built the SDK regulary (daily), to ensure there are no breaking
changes introduced to the trunk branch. However, it is rare that there
actually is a need to update to a newly built SDK on developer machines.
As the development environment will decide based on the SDK checksum
stored in the registries, whether an update is necessary, we make the
uploading new builds to the registry optional.