[DEVOPS-691] Enable hashserver in CI build
During a build, it may often be the case that the output / result of a task
might be unchanged despite changes in the task’s input values.
An example might be whitespace changes in some input C code.
In project terms, this is what we define as “equivalence”.

This commit enables the hashserver within the pipeline builds which will
reduce the number of package rebuilds due to changes in upstream
recipes. This is only relevant for development pipelines, as releases
are built from scratch anyway.

The hashserver database is populated on CI builds on the default branch
(develop) and is consumed RO on any other branch pipeline.