[DEVOPS-813] Fine-granular build config options
Previously, there where some limitations to the fine-granularity of
target configurations. For example, all target layer repositories needed
to be identical.

Not only was this a potential source of undesired side-effects, e.g.
when including a irma6r1 relevant repository in a irma6r2 build, but
this also limits scalability in regards to the project portfolio.

This change adds the necessary fine-granularity to modify any aspect of
a yocto build target.