sdk-upload: Make independent of tests
The operability of the platform application does not have any impact on
the functionality of the SDK. Thus, it does not make sense to require
passing OCI tests for SDKs to upload.

Additionally, the release-build-sdk job will never run in conjunction
with the sdk uploads, due to job run rules.