gitlab-ci: Ensure latest changes are used in test-oci
There is a bug in Gitlab CI which causes service containers not to pull
the latest available container image. This can cause issues under
certain circumstances in our test-oci gitlab job, namingly when a
test-oci job on the same branch but with a newer commit  happens be
scheduled on the same node as a previous test-oci job on the same branch
but on an older commit.

While this is unlikely to happen, use the commit short SHA in the
container tag, which ensures that the latest commit is always used
for the test job.

The bug has been reported in upstream:
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/37496