Use renovate for updating upstream meta-layers
Previuosly, we used "autoref" for updating upstream meta-layers on
development builds. While this had the benefit that we always use the
latest bug- and CVE-fixes this also had two major drawbacks:

1. Since only CI jobs on the develop branch populate the shared sstate cache
   a "badly" timed upstream change to "core" packages can drastically
   worsen the Pipeline build time on feature branches until the next
  develop CI job has been run, due to a "trickle-down" effect in build
  dependencies. This has a negative impact on developers CI experience

2. Patches developed by iris employees for thirdparty meta-layers and
   applied using the KAS "patch" feature can break the CI build on
   changes in upstream layers, especially if said patch is sent
   upstream. That meant that CI would break for all developers until the
   conflict was resolved or a manual roll-back was merged.

With this change we now set a fixed commit for upstream repositories and
use renovatebot with a custom ruleset to track for changes in
upstream repositories. Doing so, we can now test and adjust for upstream
changes before merging and postpone updates if necessary without having
to manually roll-back.

Signed-off-by: Jasper Orschulko <jasper@fancydomain.eu>