Replace meta-dependencytrack with meta-cyclonedx
The main advantage of using meta-cyclonedx is that it separates
uploading the bom to dependencytrack from the actual build step.

This makes the build itself more stable and independent of the
availability of a dependencytrack server.

The logic to upload the bom and vex file has therefore been outsourced
into its own script that runs as a follow-up job in the gitlab ci
pipeline. This also allows easy re-triggering of the upload job after a
temporary infrastructure failure without having to rerun the entire
build.