utils: Remove versioning fallback schema
The fallback schema was in place for the period in time when there were
still existing products with no release in the new versioning schema.
Since this is not the case anymore, not finding an existing previous
release means that we are dealing with a new product without any
pre-existing releases.